UI + Network 2

+++ app/src/main/java/bme/hu/sziaclientapp/network/FlightsApi.java
@@ -28,7 +28,7 @@ public interface FlightsApi {
   
   @GET("flights/{id}")
   Call<Flight> flightsIdGet(
-          @Path("id") Integer id
+          @Path("id") Long id
   );
